Make App trait generic over response size
The buffer size is an implementation detail of the dispatch
implementation.  Applications should not depend on it.  This patch makes
the App trait generic over the response size and changes the request
argument to use a slice instead of a reference to a heapless::Vec.

If applications need a minimum response size, they can still use a
const assertion to enforce it.